[UPDATE] Move createInvoice to invoice dir
